requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
